refactor(backend:platform): 收拢注册表与系统代理实现
