+++ src-tauri/src/platform/windows/mod.rs
@@ -7,6 +7,8 @@ pub(super) mod dpapi;
 mod geometry;
 pub(super) mod hotkey;
 pub(super) mod input;
+pub(super) mod proxy;
+pub(super) mod registry;
 pub(super) mod sound;
 pub(super) mod webview2;
 pub(super) mod window;
